Backend/requirements.txt
Flask==3.0.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
Werkzeug==3.0.1
